Revert "Add support to write codegen scalar functions"
This reverts commit d02882d77add4a26c5afdcb02ef0df36ecc44ea0.
