+++ presto-main/src/main/java/com/facebook/presto/sql/gen/BodyCompiler.java
@@ -13,7 +13,6 @@
  */
 package com.facebook.presto.sql.gen;
 
-import com.facebook.presto.bytecode.CallSiteBinder;
 import com.facebook.presto.bytecode.ClassDefinition;
 import com.facebook.presto.common.function.SqlFunctionProperties;
 import com.facebook.presto.spi.relation.RowExpression;
