+++ presto-main/src/main/java/com/facebook/presto/sql/gen/CachedInstanceBinder.java
@@ -13,9 +13,7 @@
  */
 package com.facebook.presto.sql.gen;
 
-import com.facebook.presto.bytecode.Binding;
 import com.facebook.presto.bytecode.BytecodeBlock;
-import com.facebook.presto.bytecode.CallSiteBinder;
 import com.facebook.presto.bytecode.ClassDefinition;
 import com.facebook.presto.bytecode.FieldDefinition;
 import com.facebook.presto.bytecode.Variable;
