+++ presto-main/src/main/java/com/facebook/presto/sql/gen/CursorProcessorCompiler.java
@@ -16,7 +16,6 @@ package com.facebook.presto.sql.gen;
 import com.facebook.airlift.log.Logger;
 import com.facebook.presto.bytecode.BytecodeBlock;
 import com.facebook.presto.bytecode.BytecodeNode;
-import com.facebook.presto.bytecode.CallSiteBinder;
 import com.facebook.presto.bytecode.ClassDefinition;
 import com.facebook.presto.bytecode.MethodDefinition;
 import com.facebook.presto.bytecode.Parameter;
