+++ presto-main/src/main/java/com/facebook/presto/sql/gen/SqlTypeBytecodeExpression.java
@@ -13,21 +13,19 @@
  */
 package com.facebook.presto.sql.gen;
 
-import com.facebook.presto.bytecode.Binding;
 import com.facebook.presto.bytecode.BytecodeNode;
-import com.facebook.presto.bytecode.CallSiteBinder;
 import com.facebook.presto.bytecode.MethodGenerationContext;
 import com.facebook.presto.bytecode.expression.BytecodeExpression;
 import com.facebook.presto.bytecode.instruction.InvokeInstruction;
 import com.facebook.presto.common.type.Type;
+import com.google.common.collect.ImmutableList;
 import io.airlift.slice.Slice;
 
 import java.lang.reflect.Method;
-import java.util.Collections;
 import java.util.List;
 
-import static com.facebook.presto.bytecode.BytecodeUtils.Bootstrap.BOOTSTRAP_METHOD;
 import static com.facebook.presto.bytecode.ParameterizedType.type;
+import static com.facebook.presto.sql.gen.Bootstrap.BOOTSTRAP_METHOD;
 import static java.util.Objects.requireNonNull;
 
 public class SqlTypeBytecodeExpression
@@ -73,7 +71,7 @@ public class SqlTypeBytecodeExpression
     @Override
     public List<BytecodeNode> getChildNodes()
     {
-        return Collections.emptyList();
+        return ImmutableList.of();
     }
 
     @Override
